agents/tutor_agent/requirements.txt
bedrock-agentcore
strands-agents
boto3
requests
pydantic

# Utilities
python-dateutil>=2.8.2
